+++ typescript/packages/subsurface-viewer/src/components/Map.tsx
@@ -1733,9 +1733,10 @@ function computeViewState(
         };
         return updateViewState(defaultCamera, boundingBox, zScale, size);
     } else {
+        const is3D = false;
+
         // If the camera is defined, use it
         if (isCameraPositionDefined) {
-            const is3D = false;
             return updateViewState(
                 cameraPosition,
                 boundingBox,
@@ -1746,6 +1747,10 @@ function computeViewState(
         }
 
         const centerOfData: Point3D = boxCenter(boundingBox);
+
+        // In 2D set camera target (centerOfData) to 0; Zoom scales world around target.
+        centerOfData[2] = 0;
+
         // if bounds are defined, use them
         if (isBoundsDefined) {
             return getViewStateFromBounds(