+++ typescript/packages/subsurface-viewer/CHANGELOG.md
@@ -1,3 +1,17 @@
+## [0.30.6](https://github.com/equinor/webviz-subsurface-components/compare/subsurface-viewer@0.30.5...subsurface-viewer@0.30.6) (2024-09-17)
+
+
+### Bug Fixes
+
+* allow initializing scaling factor value ([#2248](https://github.com/equinor/webviz-subsurface-components/issues/2248)) ([10b1901](https://github.com/equinor/webviz-subsurface-components/commit/10b190147c749bc2558bc2216fc50184a2675b17))
+
+## [0.30.5](https://github.com/equinor/webviz-subsurface-components/compare/subsurface-viewer@0.30.4...subsurface-viewer@0.30.5) (2024-09-16)
+
+
+### Bug Fixes
+
+* 2D axes labels disappears when max zoom in. [#2229](https://github.com/equinor/webviz-subsurface-components/issues/2229) ([#2230](https://github.com/equinor/webviz-subsurface-components/issues/2230)) ([95d557c](https://github.com/equinor/webviz-subsurface-components/commit/95d557c6f7023e480b7a07ba9efdb7998fee71e8))
+
 ## [0.30.4](https://github.com/equinor/webviz-subsurface-components/compare/subsurface-viewer@0.30.3...subsurface-viewer@0.30.4) (2024-09-11)